+++ src/Shared/Avalon.Network.Packets/Movement/SPlayerPositionUpdatePacket.cs
@@ -42,8 +42,8 @@ public class SPlayerPositionUpdatePacket : Packet
 [ProtoContract]
 public class SPlayerPacket
 {
-    [ProtoMember(1)] public int AccountId { get; set; }
-    [ProtoMember(2)] public int CharacterId { get; set; }
+    [ProtoMember(1)] public ulong AccountId { get; set; }
+    [ProtoMember(2)] public ulong CharacterId { get; set; }
     [ProtoMember(3)] public float PositionX { get; set; }
     [ProtoMember(4)] public float PositionY { get; set; }
     [ProtoMember(5)] public float VelocityX { get; set; }
